Create useDragAndDrop hook to DRY things up
